Big clean up to get the MK building on MacOS 10.9, upgraded all the sundry pbproj's to xcodeproj's, now compiles with latest HIP embedded within LAME - first layer of bitrot removed

+++ MusicKit/Frameworks/MusicKit/MKNoteReceiver.m
@@ -324,7 +324,7 @@ only the owner itself sends this message.
 
 - _disconnect: (MKNoteSender *) aNoteSender
 {
-    unsigned int returnedIndex;
+    NSUInteger returnedIndex;
     
     if ((returnedIndex = [noteSenders indexOfObjectIdenticalTo: aNoteSender]) != NSNotFound) {
         [noteSenders removeObjectAtIndex: returnedIndex];
